% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simset_class.R
\name{add.parameters}
\alias{add.parameters}
\title{Add parameters to a simset or parameter_set}
\usage{
add.parameters(
  simset,
  parameters,
  parameter.names = NULL,
  parameter.transformations = NULL,
  parameter.lower.bounds = -Inf,
  parameter.upper.bounds = Inf
)
}
\arguments{
\item{simset}{A \link{simset} or \link{parameter_set} to which to add parameters}

\item{parameters}{The parameters to add. If only one new parameter, can be either a vector of length simset@n.sim or length 1, or a matrix with one column and simset@n.sim rows. If more than one parameter is to be added, a matrix with one column for parameter and simset@n.sim rows}

\item{parameter.names}{The names of the new parameters being added. If NULL, takes the parameter names from the column names of parameters}

\item{parameter.transformations}{Objects of class \link{distributions::transformation}, or the names of defined transformations, according to which parameters should be smoothed when making smoothed distributions}

\item{parameter.lower.bounds, parameter.upper.bounds}{The min and max values each parameter can take. If given a scalar value, will apply the same bounds to all parameters}
}
\description{
Add parameters to a simset or parameter_set
}
