% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interface.R
\name{cache.progress}
\alias{cache.progress}
\title{Report the progress of an MCMC process running from a cache}
\usage{
cache.progress(dir)
}
\arguments{
\item{dir}{The directory where the cache is stored}
}
\description{
Report the progress of an MCMC process running from a cache
}
