% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_class.R
\name{check.merge.controls}
\alias{check.merge.controls}
\title{Check whether mcmc objects resulting from two mcmcsim_control objects can be merged}
\usage{
check.merge.controls(c1, c2, for.serial.merge = T)
}
\value{
If the two controls are merge-able, returns an empty character vector. Otherwise, returns a vector of the reasons why the two cannot be merged
}
\description{
Check whether mcmc objects resulting from two mcmcsim_control objects can be merged
}
