% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interface.R
\name{create.mcmc.cache}
\alias{create.mcmc.cache}
\title{Set up a cache for running an MCMC from}
\usage{
create.mcmc.cache(
  dir,
  control,
  n.iter,
  starting.values = NULL,
  prior.mcmc = NULL,
  cache.frequency = 1000,
  allow.overwrite.cache = F,
  merge.with.prior.mcmc = T
)
}
\arguments{
\item{dir}{The directory where the cache should be stored}

\item{cache.frequency}{How often to cache to the disk}

\item{allow.overwrite.cache}{If a previous cache is in place in dir, the function will throw an error if allow.overwrite.cache is FALSE. If allow.overwrite.cache is TRUE, it will overwrite the previously existing cache}

\item{cache.prefix}{???}
}
\description{
Set up a cache for running an MCMC from
}
