% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simset_class.R
\name{extend.simulations}
\alias{extend.simulations}
\title{Extend the simulations of a simset}
\usage{
extend.simulations(simset, fn)
}
\arguments{
\item{simset}{An object of class \link{simset}}

\item{fn}{A function that takes two arguments: sim (the prior simulation object) and parameters (the corresponding numeric vector of parameters) and returns a new simulation object}
}
\value{
An updated simset
}
\description{
Extend the simulations of a simset
}
