% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simset_class.R
\name{extract.simset.distribution}
\alias{extract.simset.distribution}
\title{Extract a distribution of simulation results}
\usage{
extract.simset.distribution(
  simset,
  fn,
  pass.to.fn = c("simulation", "parameters", "both")[1],
  ...,
  smooth = F
)
}
\arguments{
\item{simset}{A \link{simset} object}

\item{fn}{A function to be applied. If pass.to.fn=='simulation', the function's first parameters should be 'sim' (a simulation objet). If pass.to.fn=='parameters', the function's first argument should be 'parameters' (a vector of parameters). If pass.to.fn=='both', the function's first two parameters should be 'sim' and 'parameters'}

\item{pass.to.fn}{What to pass to fn as arguments}

\item{...}{Additional parameters to be passed to fn}

\item{smooth}{Whether to return a smoothed distribution over continuous parameters}
}
\description{
Gets a probability distribution over the results of a function applied to each simulation or its parameters or both
}
