% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simset_class.R
\name{extract.simset.parameter.distribution}
\alias{extract.simset.parameter.distribution}
\title{Extract the distribution of parameters in a simset}
\usage{
extract.simset.parameter.distribution(simset, smooth = F)
}
\arguments{
\item{simset}{A \link{simset} object}

\item{smooth}{Whether to return a smoothed distribution over continuous parameters}
}
\description{
Gets a probability distribution over the parameters in a simset
}
