% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simset_class.R
\name{extract.simset}
\alias{extract.simset}
\title{Extracts a simset from an object containing simulations}
\usage{
extract.simset(object, ...)
}
\arguments{
\item{object}{The object from which to extract the simset}
}
\description{
Extracts a simset from an object containing simulations
}
