% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get.rhats}
\alias{get.rhats}
\title{Calculate R-hat statistics for an MCMC object}
\usage{
get.rhats(
  mcmc,
  var.names = mcmc@var.names,
  chains = 1:mcmc@n.chains,
  additional.burn = 0,
  additional.thin = 1,
  exact.var.names = F,
  rank.normalize = F,
  sort = T
)
}
\arguments{
\item{mcmc}{A \link{mcmc.sim} object}

\item{var.names}{var.names is the names of variables to include or a set of integer indices. If exact.var.names is TRUE, matches the given strings exactly. If exact.var.names is FALSE, matches any variable name that begins with any element of var.names, where the * character is treated as a wild card}

\item{chains}{Which chains from the MCMC to use}

\item{exact.var.names}{var.names is the names of variables to include or a set of integer indices. If exact.var.names is TRUE, matches the given strings exactly. If exact.var.names is FALSE, matches any variable name that begins with any element of var.names, where the * character is treated as a wild card}

\item{rank.normalize}{Whether to rank-normalize samples before calculating the R-hat, as in Vehtari 2019}

\item{sort}{If TRUE, the rhats are sorted from greatest to least}
}
\description{
Calculates the R-hat of Gelman 2013, or the R-hat using rank-normalization as per Vehtari 2019
}
