% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is.mcmc.cache.complete}
\alias{is.mcmc.cache.complete}
\title{Check if a cached MCMC has completed sampling}
\usage{
is.mcmc.cache.complete(dir)
}
\arguments{
\item{dir}{The cache directory}
}
\value{
A boolean indicator of whether the cache is done sampling
}
\description{
Check if a cached MCMC has completed sampling
}
