% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mcmc.merge.parallel}
\alias{mcmc.merge.parallel}
\title{Merge MCMC objects in parallel}
\usage{
mcmc.merge.parallel(...)
}
\description{
Merges two or more mcmcsim objects by treating the chains in each as parallel chains
}
