% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mcmc.merge.serial}
\alias{mcmc.merge.serial}
\title{Merge MCMC objects in serial}
\usage{
mcmc.merge.serial(...)
}
\arguments{
\item{...}{One or more mcmcsim objects or lists containing only mcmcsim objects}
}
\description{
Merges two or more mcmcsim objects by treating the chains in each a serial chains (ie, extensions of the previous chains)
}
