% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_class.R
\name{mcmc.subset}
\alias{mcmc.subset}
\title{Subset an MCMC object}
\usage{
mcmc.subset(
  .Object,
  chains = 1:.Object@n.chains,
  additional.burn = 0,
  additional.thin = 1
)
}
\arguments{
\item{.Object}{The mcmc.sim object}

\item{chains}{The chains to subset}

\item{additional.burn}{Additional number of iterations to burn off the front (this represents the number of iterations AFTER the thinning which has already been done to the mcmc.sim object but BEFORE applying additional.thin)}

\item{additional.thin}{The factor by which to further thin iterations}
}
\value{
An mcmc.sim object
}
\description{
Subset an MCMC object by either burning additional iterations of the front, further thinning iterations, selecting a subset of chains, or some combination of the three
}
