% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{pair.plot}
\alias{pair.plot}
\title{Generate a pairs plot of variables from an mcmc run}
\usage{
pair.plot(
  mcmc,
  var.names = mcmc@var.names,
  chains = 1:mcmc@n.chains,
  additional.burn = 0,
  additional.thin = 1,
  exact.var.names = F
)
}
\arguments{
\item{mcmc}{A \link{mcmc.sim} object}

\item{var.names}{var.names is the names of variables to include or a set of integer indices. If exact.var.names is TRUE, matches the given strings exactly. If exact.var.names is FALSE, matches any variable name that begins with any element of var.names, where the * character is treated as a wild card}

\item{chains}{Which chains from the MCMC to use}

\item{exact.var.names}{var.names is the names of variables to include or a set of integer indices. If exact.var.names is TRUE, matches the given strings exactly. If exact.var.names is FALSE, matches any variable name that begins with any element of var.names, where the * character is treated as a wild card}
}
\description{
Generate a pairs plot of variables from an mcmc run
}
