% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{process.transformations}
\alias{process.transformations}
\title{exporting for now for debugging}
\usage{
process.transformations(
  var.names,
  transformations,
  n.var = length(var.names),
  var.names.name = "var.names"
)
}
\description{
exporting for now for debugging
}
