% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interface.R
\name{remove.mcmc.cache}
\alias{remove.mcmc.cache}
\title{Delete previously set-up cache for running MCMC}
\usage{
remove.mcmc.cache(dir, allow.remove.incomplete = F)
}
\arguments{
\item{dir}{The directory where the cache is stored}

\item{allow.remove.incomplete}{If FALSE, the function will only succeed in removing the cached files if all the MCMC chains are complete. If TRUE, removes all file regardless}
}
\description{
Delete previously set-up cache for running MCMC
}
