% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interface.R
\name{run.mcmc.from.cache}
\alias{run.mcmc.from.cache}
\title{Run MCMC from a previously set-up cache}
\usage{
run.mcmc.from.cache(
  dir,
  chains = NULL,
  update.frequency = 100,
  update.detail = c("none", "low", "med", "high")[3],
  cores = parallel::detectCores(),
  remove.cache.when.done = F,
  output.file = ""
)
}
\arguments{
\item{dir}{The directory where the cache is stored}

\item{chains}{The numbers of the chains to run. If passed NULL (the default), runs all chains}

\item{update.frequency}{How often to print updates (print an update every update.frequency iterations). If NA, no updates are printed}

\item{update.detail}{The level of detail for updates}

\item{cores}{The maximum number of cores to use to run chains in parallel}
}
\value{
An MCMC object with length(chains) number of chains
}
\description{
Runs MCMC iterations from a previously set up cache
}
