% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simset_class.R
\docType{class}
\name{simset}
\alias{simset}
\alias{simset-class}
\title{The simset class}
\description{
A class representing a weighted set of simulations
}
\section{Slots}{

\describe{
\item{\code{simulations}}{A list of simulations}

\item{\code{weights}}{A numeric weight for each simulation}

\item{\code{n.sim}}{The number of unique simulations}

\item{\code{parameters}}{A matrix of parameters for each simulation, where row i corresponds to the parameters used to run the ith simulation, and each column represents a parameter value}

\item{\code{parameter.names,n.parameters}}{The names and number of the parameters used to run simulations}
}}

