% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simset_class.R
\name{subset.simset}
\alias{subset.simset}
\title{Subset a simset}
\usage{
subset.simset(simset, indices)
}
\arguments{
\item{simset}{An object of class simset}

\item{indices}{Either numeric indices or a logical mask according to which to subset simulations}
}
\value{
A new simset object
}
\description{
Subset a simset by indexing a subset of the included simulations
}
