% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simset_class.R
\name{thin.simset}
\alias{thin.simset}
\title{Subset a simset by thinning its simulations}
\usage{
thin.simset(simset, thin)
}
\arguments{
\item{simset}{A simset object}

\item{thin}{How much to thin the simulations}
}
\value{
A thinned simset object
}
\description{
For a simset whose weights are integer values, extracts a subset of the simualtions by thinning every nth simulation (eg, if thin=3, the resulting simset keeps every third simulation). Maintains the weighting - simulations with higher weights are more likely to be kept
}
