% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{trace.plot}
\alias{trace.plot}
\title{Generate a trace plot from an mcmc run}
\usage{
trace.plot(
  mcmc,
  var.names = mcmc@control@var.names,
  chains = 1:mcmc@n.chains,
  additional.burn = 0,
  additional.thin = 1,
  exact.var.names = F,
  include.rhat = mcmc@n.chains > 1,
  ncol = NULL
)
}
\arguments{
\item{mcmc}{A \link{mcmc.sim} object}

\item{var.names, exact.var.names}{var.names is the names of variables to include or a set of integer indices. If exact.var.names is TRUE, matches the given strings exactly. If exact.var.names is FALSE, matches any variable name that begins with any element of var.names, where the * character is treated as a wild card}

\item{chains}{Which chains from the MCMC to use}

\item{include.rhat}{If TRUE, includes the rhat value in the facet header}

\item{ncol}{Number of columns in the facet_wrap for the resulting plot object}
}
\description{
Generate a trace plot from an mcmc run
}
